[Add] videoDetail, css

+++ controllers/videoController.js
@@ -46,9 +46,12 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
     // 여기있는 fileUrl, title, description은 videoDB의 속성이다.
   });
-  console.log(newVideo);
+  // console.log(newVideo);
+  req.user.videos.push(newVideo.id); // user DB의 video atribute에 추가
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id)); // id는 DB의 id
   // id는 mongoDB가 랜덤하게 만들어준다.
 };
@@ -59,7 +62,7 @@ export const videoDetail = async (req, res) => {
     params: { id },
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -72,7 +75,11 @@ export const getEditVideo = async (req, res) => {
   try {
     const video = await Video.findById(id);
     // video를 받아서 render로 통해 템플릿으로 던져준다,
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
     // rendering하는 순간 템플릿에선 video의 title과 description을 던져준다.
   } catch (error) {
     res.redirect(routes.home);
@@ -99,7 +106,13 @@ export const deleteVideo = async (req, res) => {
     params: { id },
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    // video를 받아서 render로 통해 템플릿으로 던져준다,
+    if (String(video.creator) !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
